requirements.txt
streamlit>=1.24.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.1
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
duckduckgo-search>=3.8.3
requests>=2.28.0
beautifulsoup4>=4.11.0
